core/tests/test_plugins.py
"""Tests for ADK Plugins (core/orrery_core/plugins.py)."""

from __future__ import annotations

from unittest.mock import AsyncMock, MagicMock, patch

import pytest

from orrery_core.plugins import (
    ActivityPlugin,
    AuditPlugin,
    ErrorHandlerPlugin,
    GuardrailsPlugin,
    MemoryPlugin,
    MetricsPlugin,
    ResiliencePlugin,
    default_plugins,
)
from orrery_core.rbac import Role, RolePolicy

# Fixtures for ADK mock objects


@pytest.fixture
def base_tool():
    tool = MagicMock()
    tool.name = "my_tool"
    # Mock the underlying function to handle guardrail checks
    tool.func = lambda: None
    return tool


@pytest.fixture
def tool_context():
    ctx = MagicMock()
    ctx.state = {}
    ctx.agent_name = "test_agent"
    return ctx


@pytest.fixture
def callback_context():
    ctx = MagicMock()
    ctx.state = {}
    return ctx


@pytest.fixture
def base_agent():
    agent = MagicMock()
    agent.name = "test_agent"
    return agent


# ── GuardrailsPlugin Tests ───────────────────────────────────────────


@pytest.mark.asyncio
async def test_guardrails_plugin_rbac_blocks(base_tool, tool_context):
    """Verify RBAC check blocks unauthorized tool calls."""
    policy = RolePolicy(overrides={"my_tool": Role.ADMIN}, default_role=Role.VIEWER)
    plugin = GuardrailsPlugin(role_policy=policy)

    # Mock context with viewer role (not authorized for "my_tool")
    tool_context.state["user_role"] = "viewer"

    result = await plugin.before_tool_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context
    )

    assert result is not None
    assert result["status"] == "access_denied"
    assert "Access denied" in result["message"]


@pytest.mark.asyncio
async def test_guardrails_plugin_confirm_mode_skips_gate(base_tool, tool_context):
    """In confirm mode, GuardrailsPlugin handles RBAC only — confirmation is
    handled at the agent level via before_tool_callback."""
    from orrery_core.guardrails import confirm

    @confirm("testing")
    def my_guarded_func():
        pass

    base_tool.func = my_guarded_func
    base_tool.name = "my_tool"

    plugin = GuardrailsPlugin(mode="confirm")
    tool_context.state["user_role"] = "admin"  # Authorized

    result = await plugin.before_tool_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context
    )

    # No confirmation gate at plugin level — RBAC passes, tool proceeds
    assert result is None


@pytest.mark.asyncio
async def test_guardrails_plugin_dry_run_blocks(base_tool, tool_context):
    """Verify dry_run mode still blocks guarded tools at the plugin level."""
    from orrery_core.guardrails import destructive

    @destructive("deletes data")
    def my_destructive_func():
        pass

    base_tool.func = my_destructive_func
    base_tool.name = "my_tool"

    plugin = GuardrailsPlugin(mode="dry_run")
    tool_context.state["user_role"] = "admin"

    result = await plugin.before_tool_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context
    )

    assert result is not None
    assert result["status"] == "dry_run"


# ── ResiliencePlugin Tests ───────────────────────────────────────────


@pytest.mark.asyncio
async def test_resilience_plugin_opens_circuit(base_tool, tool_context):
    """Verify ResiliencePlugin blocks calls when circuit is open."""
    plugin = ResiliencePlugin(failure_threshold=1)

    # Trigger a failure to open circuit
    await plugin.on_tool_error_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context, error=Exception("Fail")
    )

    # Next call should be blocked by before_tool_callback
    result = await plugin.before_tool_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context
    )

    assert result is not None
    assert "temporarily unavailable" in result["message"]


# ── MetricsPlugin Tests ──────────────────────────────────────────────


@pytest.mark.asyncio
async def test_metrics_plugin_tracks_calls(base_tool, tool_context):
    """Verify MetricsPlugin calls the underlying collector."""
    plugin = MetricsPlugin()

    with patch.object(plugin, "_before", wraps=plugin._before) as mock_before:
        await plugin.before_tool_callback(tool=base_tool, tool_args={}, tool_context=tool_context)
        mock_before.assert_called_once()


# ── AuditPlugin Tests ────────────────────────────────────────────────


@pytest.mark.asyncio
async def test_audit_plugin_logs_call(base_tool, tool_context):
    """Verify AuditPlugin calls the audit logger."""
    plugin = AuditPlugin()

    with patch.object(plugin, "_callback", wraps=plugin._callback) as mock_audit:
        await plugin.after_tool_callback(
            tool=base_tool, tool_args={}, tool_context=tool_context, result={"status": "success"}
        )
        mock_audit.assert_called_once()


# ── ActivityPlugin Tests ─────────────────────────────────────────────


@pytest.mark.asyncio
async def test_activity_plugin_updates_state(base_tool, tool_context):
    """Verify ActivityPlugin updates session activity log."""
    plugin = ActivityPlugin()

    await plugin.after_tool_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context, result={"status": "success"}
    )

    assert "session_log" in tool_context.state
    assert len(tool_context.state["session_log"]) == 1
    assert tool_context.state["session_log"][0]["operation"] == "my_tool"


# ── ErrorHandlerPlugin Tests ─────────────────────────────────────────


@pytest.mark.asyncio
async def test_error_handler_plugin_suppresses_tool_error(base_tool, tool_context):
    """Verify ErrorHandlerPlugin returns a structured error dict."""
    plugin = ErrorHandlerPlugin()

    result = await plugin.on_tool_error_callback(
        tool=base_tool, tool_args={}, tool_context=tool_context, error=ValueError("Invalid arg")
    )

    assert isinstance(result, dict)
    assert result["status"] == "error"
    assert "Invalid arg" in result["message"]


@pytest.mark.asyncio
async def test_error_handler_plugin_suppresses_model_error(callback_context):
    """Verify ErrorHandlerPlugin returns a LlmResponse on model error."""
    plugin = ErrorHandlerPlugin()
    llm_request = MagicMock()

    result = await plugin.on_model_error_callback(
        callback_context=callback_context, llm_request=llm_request, error=Exception("Model timeout")
    )

    from google.adk.models.llm_response import LlmResponse

    assert isinstance(result, LlmResponse)
    assert result.content is not None
    assert result.content.parts is not None
    assert len(result.content.parts) > 0
    text = result.content.parts[0].text
    assert text is not None
    assert "Model timeout" in text


# ── MemoryPlugin Tests ──────────────────────────────────────────────


@pytest.fixture
def memory_callback_context():
    """CallbackContext mock with invocation_context for memory tests."""
    ctx = MagicMock()
    ctx.state = {}
    # Root agent
    ctx._invocation_context.agent.name = "root_agent"
    # Session with enough events
    ctx._invocation_context.session.events = [MagicMock() for _ in range(6)]
    ctx._invocation_context.session.id = "sess-123"
    # Memory service mock
    ctx._invocation_context.memory_service = MagicMock()
    ctx._invocation_context.memory_service.add_session_to_memory = AsyncMock()
    return ctx


@pytest.fixture
def root_agent():
    agent = MagicMock()
    agent.name = "root_agent"
    return agent


@pytest.mark.asyncio
async def test_memory_plugin_saves_root_session(memory_callback_context, root_agent):
    """MemoryPlugin saves session when root agent finishes with enough events."""
    plugin = MemoryPlugin(min_events=4)

    await plugin.after_agent_callback(agent=root_agent, callback_context=memory_callback_context)

    memory_callback_context._invocation_context.memory_service.add_session_to_memory.assert_called_once_with(
        memory_callback_context._invocation_context.session
    )


@pytest.mark.asyncio
async def test_memory_plugin_skips_trivial_session(memory_callback_context, root_agent):
    """MemoryPlugin skips sessions with fewer events than min_events."""
    memory_callback_context._invocation_context.session.events = [MagicMock(), MagicMock()]
    plugin = MemoryPlugin(min_events=4)

    await plugin.after_agent_callback(agent=root_agent, callback_context=memory_callback_context)

    memory_callback_context._invocation_context.memory_service.add_session_to_memory.assert_not_called()


@pytest.mark.asyncio
async def test_memory_plugin_skips_sub_agent(memory_callback_context):
    """MemoryPlugin only fires for the root agent."""
    sub_agent = MagicMock()
    sub_agent.name = "sub_agent"
    plugin = MemoryPlugin(min_events=4)

    await plugin.after_agent_callback(agent=sub_agent, callback_context=memory_callback_context)

    memory_callback_context._invocation_context.memory_service.add_session_to_memory.assert_not_called()


@pytest.mark.asyncio
async def test_memory_plugin_skips_no_memory_service(root_agent):
    """MemoryPlugin handles missing memory_service gracefully."""
    ctx = MagicMock()
    ctx._invocation_context.agent.name = "root_agent"
    ctx._invocation_context.memory_service = None
    ctx._invocation_context.session.events = [MagicMock() for _ in range(6)]
    plugin = MemoryPlugin(min_events=4)

    result = await plugin.after_agent_callback(agent=root_agent, callback_context=ctx)

    assert result is None


# ── Factory Tests ────────────────────────────────────────────────────


def test_default_plugins_composition():
    """Verify default_plugins returns the expected list and order."""
    plugins = default_plugins()

    expected_names = ["guardrails", "resilience", "metrics", "audit", "activity", "error_handler"]

    # ADK BasePlugin has a .name property
    plugin_names = [p.name for p in plugins]
    assert plugin_names == expected_names

    # Verify ErrorHandlerPlugin is last
    assert isinstance(plugins[-1], ErrorHandlerPlugin)


def test_default_plugins_with_memory():
    """Verify enable_memory adds MemoryPlugin before ErrorHandlerPlugin."""
    plugins = default_plugins(enable_memory=True)

    expected_names = [
        "guardrails",
        "resilience",
        "metrics",
        "audit",
        "activity",
        "memory",
        "error_handler",
    ]
    plugin_names = [p.name for p in plugins]
    assert plugin_names == expected_names

    # MemoryPlugin is present and ErrorHandlerPlugin is still last
    assert isinstance(plugins[-2], MemoryPlugin)
    assert isinstance(plugins[-1], ErrorHandlerPlugin)
