core/pyproject.toml
[project]
name = "orrery-core"
version = "0.1.2"
description = "Shared utilities for AI agents"
requires-python = ">=3.11"
dependencies = [
    "google-adk[eval]>=1.27.1",
    "litellm>=1.60.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "prometheus-client>=0.21.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
# Install with: uv sync --extra postgres
# Enables PostgreSQL-backed session storage for multi-instance deployments.
postgres = [
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.30.0",
]

[tool.setuptools.packages.find]
include = ["orrery_core*"]
